requirements.txt
annotated-types==0.7.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.3
flask-openapi3==3.1.3
idna==3.7
importlib-metadata==8.0.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
pydantic==2.8.2
pydantic-core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
typing-extensions==4.12.2
urllib3==2.2.2
werkzeug==3.0.3
zipp==3.19.2
